// apps/web/src/api/generated/model/organizationMembership.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Starter API
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationRole } from './organizationRole';

export interface OrganizationMembership {
  id: string;
  organizationId: string;
  userId: string;
  email?: string;
  role: OrganizationRole;
  createdAt: string;
  updatedAt: string;
}